Fragmantation option added to redirect The isn_uart bugfix on timeout In the usbfs removed possibility of fragments

+++ c/include/isn_trans.h
@@ -48,9 +48,9 @@ typedef struct {
     isn_driver_t* child;
 
     uint8_t port;
-    uint16_t rx_counter;
-    uint16_t tx_counter;
-    uint16_t tx_frame_counter;
+    size_t rx_counter;
+    size_t tx_counter;
+    size_t tx_frame_counter;
 }
 isn_trans_t;
